backend: more practical solution for BigInteger issue

+++ src/main/kotlin/mnu/repository/DistrictHouseRepository.kt
@@ -7,5 +7,5 @@ import java.util.*
 
 interface DistrictHouseRepository : JpaRepository<DistrictHouse, Long> {
     @Query("select dt.id from district_houses dt;", nativeQuery = true)
-    fun getAllIds(): List<BigInteger>
+    fun getAllIds(): List<Long>
 }
